requirements.txt
asgiref==3.4.1
boto3==1.20.15
botocore==1.23.15
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.4
django-anymail==8.4
django-cors-headers==3.8.0
django-nested-admin==3.3.3
django-render-block==0.8.1
django-storages==1.12.3
django-templated-email==3.0.0
django-templated-mail==1.1.1
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-extra-fields==3.2.1
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.3.2
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python-decouple==3.4
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
